Review : Extract ColorItem and color list to file
